requirements.txt
openai>=1.0.0
requests>=2.28.0
Pillow>=9.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
flask>=2.0.0
